Fix for down migrations

+++ database/migrations/2015_07_08_114333_simplify_tasks.php
@@ -54,9 +54,11 @@ class SimplifyTasks extends Migration
             $table->integer('break_duration')->nullable();
         });
 
-        Schema::table('users', function ($table) {
-            $table->dropColumn('dark_mode');
-        });
+        if (Schema::hasColumn('accounts', 'dark_mode')) {
+            Schema::table('users', function ($table) {
+                $table->dropColumn('dark_mode');
+            });
+        }
         Schema::table('users', function ($table) {
             $table->integer('theme_id')->nullable();
         });
